+++ R/print_html.R
@@ -58,9 +58,9 @@ print_html.estimate_grouplevel <- print_html.estimate_contrasts
 # we allow exporting HTML format based on "gt" or "tinytable"
 .check_format_backend <- function(...) {
   dots <- list(...)
-  if (is.null(dots) || !identical(dots$backend, "tt")) {
-    "html"
-  } else {
+  if (identical(dots$backend, "tt")) {
     "tt"
+  } else {
+    "html"
   }
 }